pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "phazr"
version = "1.0.0"
description = "A modern DAG-based orchestration framework for development workflows"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Team", email = "team@example.com"}
]
keywords = ["dag", "orchestration", "workflow", "automation", "phases", "devops", "pipeline"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Build Tools",
    "Topic :: System :: Systems Administration",
]

dependencies = [
    "pydantic>=2.0",
    "click>=8.0",
    "pyyaml>=6.0",
    "aiohttp>=3.8",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "black>=22.0",
    "mypy>=1.0",
    "ruff>=0.1",
]

[project.scripts]
phazr = "phazr.cli:main"

[project.urls]
"Homepage" = "https://github.com/yourorg/phazr"
"Bug Tracker" = "https://github.com/yourorg/phazr/issues"
"Documentation" = "https://github.com/yourorg/phazr#readme"

[tool.setuptools.packages.find]
where = ["."]